Move suffix to avoid circular references
Also move make_nods to DMSLevel2b for more generic handling.

WIP #2127
