+++ jwst/associations/lib/rules_level2_base.py
@@ -3,13 +3,13 @@ import copy
 import logging
 from os.path import (
     basename,
+    split,
     splitext
 )
 import re
 
 from jwst.associations import (
     Association,
-    AssociationRegistry,
     libpath
 )
 from jwst.associations.registry import RegistryMarker
@@ -29,6 +29,7 @@ from jwst.associations.lib.dms_base import (
 )
 from jwst.associations.lib.rules_level3_base import _EMPTY
 from jwst.associations.lib.rules_level3_base import Utility as Utility_Level3
+from jwst.lib.suffix import remove_suffix
 
 # Configure logging
 logger = logging.getLogger(__name__)
@@ -351,10 +352,70 @@ class DMSLevel2bBase(DMSBaseMixin, Association):
         # fail.
         return False
 
+    def make_nod_asns(self):
+        """Make background nod Associations
+
+        NIRSpec MSA can be nodded, such that the object
+        is in a different position in the slitlet.
+        The association creation simply groups these all together
+        as a single association, all exposures marked as `science`.
+        When complete, this method will create separate associations
+        each exposure becoming the single science exposure, and the
+        other exposures then become `background`.
+
+        Returns
+        -------
+        associations: [association[, ...]]
+            List of new associations to be used in place of
+            the current one.
+        """
+
+        for product in self['products']:
+            members = product['members']
+
+            # Split out the science vs. non-science
+            # The non-science exposures will get attached
+            # to every resulting association.
+            science_exps = [
+                member
+                for member in members
+                if member['exptype'] == 'science'
+            ]
+            nonscience_exps = [
+                member
+                for member in members
+                if member['exptype'] != 'science'
+            ]
+
+            # Create new associations for each science, using
+            # the other science as background.
+            results = []
+            for science_exp in science_exps:
+                asn = copy.deepcopy(self)
+                asn.data['products'] = None
+
+                product_name = remove_suffix(
+                    splitext(split(science_exp['expname'])[1])[0]
+                )[0]
+                asn.new_product(product_name)
+                new_members = asn.current_product['members']
+                new_members.append(science_exp)
+
+                for other_science in science_exps:
+                    if other_science['expname'] != science_exp['expname']:
+                        now_background = copy.copy(other_science)
+                        now_background['exptype'] = 'background'
+                        new_members.append(now_background)
+
+                new_members += nonscience_exps
+                results.append(asn)
+
+            return results
+
     def __repr__(self):
         try:
             file_name, json_repr = self.ioregistry['json'].dump(self)
-        except:
+        except Exception:
             return str(self.__class__)
         return json_repr
