+++ jwst/lib/suffix.py
